+++ src/SimpleMail
@@ -27,5 +27,5 @@
 #include "server.h"
 #include "serverreply.h"
 #ifdef WITH_SMIME
-#include "smime.h"
+#include "smimepart.h"
 #endif
